requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.1
requests==2.31.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
jupyter==1.0.0
ultralytics>=8.0.0
opencv-python
paho-mqtt
fastapi
uvicorn[standard]
sqlalchemy
alembic
streamlit
plotly
pandas
python-multipart
# Integração API .NET
requests>=2.31.0
aiohttp>=3.8.0

